Move permission calculation to a util class, so it can be used in unit tests
